modularize minimum
