+++ tfjs-core/src/ops/binary_ops.ts
@@ -326,56 +326,7 @@ function modStrict_<T extends Tensor>(a: T|TensorLike, b: T|TensorLike): T {
   return $a.mod($b);
 }
 
-/**
- * Returns the min of a and b (`a < b ? a : b`) element-wise.
- * Supports broadcasting.
- *
- * We also expose `minimumStrict` which has the same signature as this op and
- * asserts that `a` and `b` are the same shape (does not broadcast).
- *
- * ```js
- * const a = tf.tensor1d([1, 4, 3, 16]);
- * const b = tf.tensor1d([1, 2, 9, 4]);
- *
- * a.minimum(b).print();  // or tf.minimum(a, b)
- * ```
- *
- * ```js
- * // Broadcast minimum a with b.
- * const a = tf.tensor1d([2, 4, 6, 8]);
- * const b = tf.scalar(5);
- *
- * a.minimum(b).print();  // or tf.minimum(a, b)
- * ```
- *
- * @param a The first tensor.
- * @param b The second tensor. Must have the same type as `a`.
- */
-/** @doc {heading: 'Operations', subheading: 'Arithmetic'} */
-function minimum_<T extends Tensor>(
-    a: Tensor|TensorLike, b: Tensor|TensorLike): T {
-  let $a = convertToTensor(a, 'a', 'minimum');
-  let $b = convertToTensor(b, 'b', 'minimum');
-  [$a, $b] = makeTypesMatch($a, $b);
 
-  if ($a.dtype === 'bool') {
-    $a = $a.toInt();
-    $b = $b.toInt();
-  }
-
-  broadcast_util.assertAndGetBroadcastShape($a.shape, $b.shape);
-  const der = (dy: Tensor, saved: Tensor[]) => {
-    const [$a, $b] = saved;
-    const derA = () => dy.mul($a.lessEqual($b).toFloat());
-    const derB = () => dy.mul($a.greater($b).toFloat());
-    return {a: derA, b: derB};
-  };
-  return ENGINE.runKernelFunc((backend, save) => {
-    const res = backend.minimum($a, $b);
-    save([$a, $b]);
-    return res;
-  }, {a: $a, b: $b}, der, 'Minimum') as T;
-}
 
 /**
  * @deprecated
@@ -496,7 +447,6 @@ export const atan2 = op({atan2_});
 export const divStrict = op({divStrict_});
 export const floorDiv = op({floorDiv_});
 export const maximumStrict = op({maximumStrict_});
-export const minimum = op({minimum_});
 export const minimumStrict = op({minimumStrict_});
 export const mod = op({mod_});
 export const modStrict = op({modStrict_});
