+++ tfjs-core/src/ops/loss_ops.ts
@@ -22,7 +22,7 @@ import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 import {assertShapesMatch} from '../util';
 import {expandShapeToKeepDim} from './axis_util';
-import {minimum} from './binary_ops';
+import {minimum} from './minimum';
 import {op} from './operation';
 import {ones, scalar} from './tensor_ops';
