+++ tfjs-core/src/ops/ops.ts
@@ -57,6 +57,7 @@ export {maxPool} from './max_pool';
 export {maxPool3d} from './max_pool_3d';
 export {maxPoolWithArgmax} from './max_pool_with_argmax';
 export {maximum} from './maximum';
+export {minimum} from './minimum';
 export {multinomial} from './multinomial';
 export {notEqual} from './not_equal';
 export {oneHot} from './one_hot';
