+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -19,8 +19,6 @@ import './avg_pool';
 import './batchnorm';
 import './broadcast_to';
 import './batch_to_space_nd';
-import './max';
-import './maximum';
 import './concat';
 import './conv1d';
 import './conv2d';
@@ -39,7 +37,10 @@ import './less';
 import './less_equal';
 import './local_response_normalization';
 import './mat_mul';
+import './max';
 import './max_pool';
+import './maximum';
+import './minimum';
 import './not_equal';
 import './one_hot';
 import './pad';
