+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -50,6 +50,7 @@ const CHAINED_OPS = [
   'max',
   'maximum',
   'maxPool',
+  'minimum',
   'notEqual',
   'oneHot',
   'pad',
