+++ tfjs-core/src/register_all_gradients.ts
@@ -36,6 +36,7 @@ import {maxGradConfig} from './gradients/Max_grad';
 import {maximumGradConfig} from './gradients/Maximum_grad';
 import {maxPool3DGradConfig} from './gradients/MaxPool3D_grad';
 import {maxPoolGradConfig} from './gradients/MaxPool_grad';
+import {minimumGradConfig} from './gradients/Minimum_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {powGradConfig} from './gradients/Pow_grad';
@@ -79,6 +80,7 @@ const gradConfigs: GradConfig[] = [
   maximumGradConfig,
   maxPoolGradConfig,
   maxPool3DGradConfig,
+  minimumGradConfig,
   oneHotGradConfig,
   padV2GradConfig,
   powGradConfig,
